Fixed the BD_test sample irr.py. But also wrecked it again.
I converted the code to the newer syntax, and also
added stuff that seemed necessary to make it do
what it is supposed to.

Also, I've added the option to switch between the
eGFRD simulator and Brownian dynamics via bash input
arguments.

When you run it as follows:
$ python irr.py out.txt 1 1 GF
there are no problems.

However if you make it run more
than once, it does give problems
(e.g.:
$ python irr.py out.txt 1 2 GF
)

This is particularly curious because
the 2 means that a function is called
twice. But the arguments given to the
function don't change. Thus it shouldn't
matter if the simulation is performed
once or twice.
Somehow it does however.

Anyone?

For more comments see the code itself.

	modified:   irr.py