Got samples/rebind/run.py running. HOWEVER, since it made use of some functions that are no longer in the code, I had to come up with some sort of substitutes. This code doesn't function properly in all cases. Namely: * The function EGFRDSimulator.distance_between_particles() doesn't exist anymore. Instead I get particle all particle IDs of a certain type. Then I simply take the first particle in the list to calculate the distance between hopefully the right particles. This thus only works when there are only two particles.
I should really look into what we want to measure before fixing this.
A possible solution would be to just save the ID of the particle
that was placed initially with place_particle.

* The function clear_volume gives the IDs of domains in a certain
neighboorhood. It bursted these domains to get them up to the
current time. But this leads to the problem that a particle might
belong to a domain in the to be cleared area, but it's actual
position is not in the cleared area. This might lead to an infinite
loop.