Fixed "administrative" bug in GreensFunction2DRadAbs which made it function incorrectly.
Had to do with bookkeeping of the location of the last root.
Scanning for the roots makes it impossible to calculate the xth
root without knowing all previous roots i<x.
The algorithm previously incorrectly did assume that this was
possible. Now fixed.

When root table is expanded, all roots are calculated for the
length of the table. This perhaps leaves room for optimization,
because roots are sometimes calculated based upon estimated #
terms necessary.

MW.