Implement function that returns true if interction is allowed.
The function allows_interaction_from was implemented as a method to
structure.

It receives a position and returns "true" if an interaction with
the structure (be it Sphere, Cylinder, Disk or Plane) is allowed.

This can be used to define "valid" postiions from which particles
can interact with the different shapes.