Fix an assert in MixedPair2D3D that fails if 2D and 3D diff. const. are of comparable magnitude
The assert itself is not a necessary condition to be met. It checked for
whether the z-rescaled IV (math.) domain fits into the cyl. shell of the
mixed pair. This is however not necessary because lengths will be transformed
back into the shell via dividing by the scaling factor in the end.