Fix transition pair shell construction problem
The shells have to be constructed at the "real" position of the CoM,
i.e. the one with applied boundaries and projections.

In contrast, the calculations have all to be performed using the
CoM as calculated in one of the two structures (structure1).

That was not the case and is fixed by this commit.

Yet to be cleaned up after testing.