Introduce a new ('fake') interaction domain that encloses a particle at the interface between a cyl. and a plane
This is also defining the correpsonding test shell.

Even though the particle actually does not go to the cylinder,
we have to treat this case as an interaction of a plane-bound particle
with the closeby cylinder. Otherwise the current framework will not
ignore the cylinder, failing to construct a membrane-bound domain.

Since we want to construct these domains only in the situation in
which a particle just arrived to the plane, coming from the touching
cylinder, we check explicitly whether the particle is below the
(orthogonal) cylinder before we construct it.

This prevents construction attempts if particles are only close to
the cylinder. In that case, we want to construct regular plane
singles or other domains.