WIP on newBDpropagator
The problem is that we get ridiculously high rates now because the
interface particle is on the plane and the rate is rescaled with
the wrong dimension factor for reactions with particles coming
from the rod.

We have to come up with a smarter way to deal with this here.