Ignore structure overlap checks for static particles in BD mode
This is a (arguably ugly) workaround to ignore the cylinder touching
a disk at the interface between plane and cylinder. Since for now all
static particles are located on disks, they have to ignore the associated
cylinders.

Of course, checking whether the particle is immobile is just a secondary
criterion to filter out these cases. It will do the job for now, but
on the long term we want a better solution (i.e. explicitly find the
closest cylinder!)