Improve some simulator consistency checks + increase allowed no. of steps in dt~0 limit to 10^5
In long simulations with significant amount of BD steps we may
get longer chains of short updates which the code then will mark
as dt~0 steps and fail very soon without reason.

On the other hand 10^5 BD steps are not expensive, so we can
keep on checking longer before we bail out a running simulation.