Fix creation of 'CylindricalSurfacePlanarSurfaceIntermediateSingletestShell'
So far it was not even created because of a standard check that discards
creation of an interaction shell when the particle already is in the
surface that it tries to interact with.

The latter, however, is a precondition for forming the abovementioned
shell, so we needed to make an exception!