Small changes to allow for correct resetting of BD_ONLY_FLAG upon reloading and recording of total simulated time
This is to avoid that "BD only" simulations will not be carried out
correctly when there are load-save interrupts on the way.

Also, we now allow to specify whether the last time of the load file
should be used as a starting time or not, and to monitor the total
simulated time using the given instance of the simulator.